+++ R/topic.R
@@ -4,14 +4,18 @@
 #' `select_priorities()`, or text vector.
 #' For data frames, function will search for "text" variable.
 #' For priorities data frame function will search for "priorities" variable.
+#' If missing, opens the webpage containing the political topics codebook.
 #' @param dictionary The dictionary of 20 major political topics from the
 #' Comparative Agendas Project (Jones et al., 2023) is used by default.
 #' Users can also declare a custom dictionary as a vector or a list.
 #' If users declare a vector, each element is treated as a independent topic.
 #' If users declare a list of subjects and related terms, function understands
 #' names as topic and words as terms.
+#' For more information on how the CAP topics were adapted, please run
+#' `gather_topics()` to access the political topics codebook.
 #' @import dplyr
 #' @importFrom tidyr unite
+#' @importFrom textstem lemmatize_strings
 #' @return A list of topics present in each text separated by comma.
 #' @examples
 #' \donttest{
@@ -21,20 +25,25 @@
 #' gather_topics(US_News_Conferences_1960_1980[1:5, 3],
 #'               dictionary = list("military" = c("military", "gun", "war"),
 #'                                 "development" = c("development", "interest rate", "banks")))
+#' #summary(gather_topics(US_News_Conferences_1960_1980[1:5, 3]))
+#' #plot(gather_topics(US_News_Conferences_1960_1980[1:5, 3],
+#' #                   dictionary = c("military", "development")))
 #' }
 #' @export
 gather_topics <- function(.data, dictionary = "CAP") {
   Words <- topics <- NULL
+  # tries to open topic codebook if no argument is declared
+  if (missing(.data)) open_codebook(codebook = "topic")
   # get text variable
   if (inherits(.data, "priorities")) {
-    text <- stats::na.omit(.clean_token(getElement(.data, "priorities")))
+    text <- stats::na.omit(textstem::lemmatize_strings(getElement(.data, "priorities")))
   } else if (inherits(.data, "data.frame")) {
-    text <- .clean_token(getElement(.data, "text"))
-  } else text <- .clean_token(.data)
+    text <- textstem::lemmatize_strings(getElement(.data, "text"))
+  } else text <- textstem::lemmatize_strings(.data)
   # get dictionary
   if (any(dictionary ==  "CAP")) {
     dictionary <- CAP_topics %>%
-      dplyr::mutate(Words = stringr::str_replace_all(.clean_token(Words),
+      dplyr::mutate(Words = stringr::str_replace_all(textstem::lemmatize_strings(Words),
                                                      ", ", "\\\\b|\\\\b"))
     subjects <- dictionary$Words
     names(subjects) <- dictionary$Topic
@@ -72,6 +81,7 @@ gather_topics <- function(.data, dictionary = "CAP") {
 #' @import quanteda
 #' @import dplyr
 #' @importFrom stringr str_detect str_remove_all
+#' @importFrom textstem lemmatize_strings
 #' @return A list of related terms to each of the topics declared in dictionary.
 #' @details This function relies on keyword assisted topic models implemented
 #' in the `\{keyATM\}` package to find related words based on the topics
@@ -81,16 +91,23 @@ gather_topics <- function(.data, dictionary = "CAP") {
 #' “Keyword-Assisted Topic Models.”
 #' _American Journal of Political Science_, 68(2): 730-750.
 #' \doi{10.1111/ajps.12779}
+#' @examples
+#' #gather_related_terms(US_News_Conferences_1960_1980[1:5, 3], dictionary = "CAP")
+#' #gather_related_terms(US_News_Conferences_1960_1980[1:5, 3],
+#' #                     dictionary = c("military", "development"))
+#' #gather_related_terms(US_News_Conferences_1960_1980[1:5, 3],
+#' #                     dictionary = list("military" = c("military", "gun", "war"),
+#' #                                       "development" = c("development", "interest rate", "banks")))
 #' @export
 gather_related_terms <- function(.data, dictionary) {
   Words <- NULL
   thisRequires("keyATM")
   # get text variable
   if (inherits(.data, "priorities")) {
-    text <- stats::na.omit(.clean_token(getElement(.data, "priorities")))
+    text <- stats::na.omit(textstem::lemmatize_strings(getElement(.data, "priorities")))
   } else if (inherits(.data, "data.frame")) {
-    text <- .clean_token(getElement(.data, "text"))
-  } else text <- .clean_token(.data)
+    text <- textstem::lemmatize_strings(getElement(.data, "text"))
+  } else text <- textstem::lemmatize_strings(.data)
   # check dictionary
   if (any(dictionary ==  "CAP")) {
     subjects <- CAP_topics %>%